content module now lazy loaded
