+++ src/app/app.component.html
@@ -1,3 +1,1 @@
-<main class="main">
-  <router-outlet></router-outlet>
-</main>
+<router-outlet></router-outlet>
