+++ src/app/auth/components/login.component.scss
@@ -1,7 +1,3 @@
-md-card {
-  margin-top: 50px;
-}
-
 md-select {
   margin-top: -5px;
 }
