+++ src/app/content/components/category-list/category-list.component.html
@@ -1,0 +1,26 @@
+<div fxLayout>
+  <span fxFlex></span>
+  <app-add-category-form (add)="addCategory($event)"></app-add-category-form>
+</div>
+
+<div>
+  <app-category *ngFor="let c of categories$ | async"
+    [category]="c"
+    [isExpanded]="(expanded$ | async).indexOf(c.id) > -1"
+    (update)="updateCategory($event)"
+    (remove)="removeCategory($event)"
+    (addItem)="addItem($event)"
+    (updateItem)="updateItem($event)"
+    (removeItem)="removeItem($event)"
+    (toggleCategoryExpanded)="toggleCategoryExpanded($event)">
+  </app-category>
+</div>
+
+<div>
+  <app-item fxLayout fxLayoutAlign="start center"
+    *ngFor="let item of itemsWithoutCategory$ | async"
+    [item]="item"
+    (updateItem)="updateItem($event)"
+    (removeItem)="removeItem($event)">
+  </app-item>
+</div>
