+++ src/app/content/components/item-form/item-form.component.html
@@ -2,7 +2,7 @@
 <form novalidate [formGroup]="saveItemForm" (ngSubmit)="saveItem()" fxLayout="column" fxLayoutGap="10px">
   <div>
     <md-input-container>
-      <input md-input 
+      <input mdInput 
         formControlName="name"
         placeholder="Name"
         required
@@ -15,7 +15,7 @@
     </md-input-container>
 
     <md-input-container>
-      <input md-input 
+      <input mdInput 
         formControlName="url"
         placeholder="URL"
         autocomplete="off"
@@ -26,7 +26,7 @@
     </md-input-container>
 
     <md-input-container>
-      <input md-input 
+      <input mdInput 
         formControlName="account"
         placeholder="Account"
         autocomplete="off"
@@ -36,7 +36,7 @@
 
   <div>
     <md-input-container>
-      <input md-input 
+      <input mdInput 
         formControlName="username"
         placeholder="Username"
         autocomplete="off"
@@ -44,7 +44,7 @@
     </md-input-container>
 
     <md-input-container>
-      <input md-input 
+      <input mdInput 
         formControlName="password"
         placeholder="Password"
         autocomplete="off"
@@ -54,7 +54,7 @@
 
   <div fxLayout>
     <md-input-container fxFlex>
-      <textarea md-input 
+      <textarea mdInput 
         formControlName="blob"
         placeholder="Blob"
         autocomplete="off"
