+++ src/app/content/content.component.html
@@ -1,43 +1,18 @@
 <app-header
   [loggedIn]="auth.isLoggedIn"
-  (logout)="auth.logout()"
+  (logout)="logout()"
   (secretPhraseChange)="secretPhraseChange($event)"
   ></app-header>
 
-
 <div fxLayout fxLayoutAlign="center start">
   <div fxFlex="800px" fxFlex.xs="100">
+
     <md-progress-bar 
       *ngIf="loading$ | async"
       color="accent"
       mode="indeterminate">
     </md-progress-bar>
 
-    <div fxLayout>
-      <span fxFlex></span>
-      <app-add-category-form (add)="addCategory($event)"></app-add-category-form>
-    </div>
-
-    <div>
-      <app-category *ngFor="let c of categories$ | async"
-        [category]="c"
-        [isExpanded]="(expanded$ | async).indexOf(c.id) > -1"
-        (update)="updateCategory($event)"
-        (remove)="removeCategory($event)"
-        (addItem)="addItem($event)"
-        (updateItem)="updateItem($event)"
-        (removeItem)="removeItem($event)"
-        (toggleCategoryExpanded)="toggleCategoryExpanded($event)">
-      </app-category>
-    </div>
-
-    <div>
-      <app-item fxLayout fxLayoutAlign="start center"
-        *ngFor="let item of itemsWithoutCategory$ | async"
-        [item]="item"
-        (updateItem)="updateItem($event)"
-        (removeItem)="removeItem($event)">
-      </app-item>
-    </div>
+    <app-category-list></app-category-list>
   </div>
 </div>
