+++ src/app/content/content.component.ts
@@ -1,16 +1,12 @@
 import 'rxjs/add/operator/let';
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+
+import { Auth } from '../shared/auth-services';
 import { Store } from '@ngrx/store';
 import * as fromRoot from './store/reducers';
 import * as category from './store/actions/categories';
-import * as item from './store/actions/items';
-import {
-  ICategory, IItem, ICategoryItems, IUpdateCategoryPayload,
-  IRemoveCategoryPayload, IUpdateItemPayload
-} from './models';
-
 
 @Component({
   selector: 'app-content',
@@ -18,57 +14,26 @@ import {
   styleUrls: ['./content.component.scss']
 })
 export class ContentComponent implements OnInit {
-  categories$: Observable<ICategoryItems[]>;
   loading$: Observable<boolean>;
-  itemsWithoutCategory$: Observable<IItem[]>;
-  expanded$ = new BehaviorSubject<number[]>([]);
-  private _expanded: number[] = [];
-
-  constructor(public store: Store<fromRoot.State>) { }
+  
+  constructor(
+    public store: Store<fromRoot.State>,
+    public auth: Auth,
+    private router: Router
+  ) { }
 
   ngOnInit() {
-    this.categories$ = this.store.let(fromRoot.getCategoryItems);
-    this.itemsWithoutCategory$ = this.store.let(fromRoot.getItemsWithoutCategory);
     this.loading$ = this.store.let(fromRoot.getLoading);
   }
-
-  addCategory(name: string) {
-    this.store.dispatch(new category.AddCategoryAction(name) as any);
+  
+  secretPhraseChange(secret: string) {
+    this.store.dispatch(new category.SecretPhraseChangeAction(secret) as any);
   }
 
-  removeCategory(info: IRemoveCategoryPayload) {
-    if (info.cascade) {
-      this.store.dispatch(new item.RemoveItemsFromCategory(info.category.id) as any);
-    } else {
-      this.store.dispatch(new item.SetItemsCategoryByCategoryId({
-        fromCategoryId: info.category.id,
-        toCategoryId: undefined
-      }) as any);
-    }
-    this.store.dispatch(new category.RemoveCategoryAction(info) as any);
-    this.toggleCategoryExpanded(info.category.id);
-  }
-
-  updateCategory(info: IUpdateCategoryPayload) {
-    this.store.dispatch(new category.EditCategoryAction(info) as any);
-  }
-
-  addItem(itemModel: IItem) {
-    this.store.dispatch(new item.AddItemAction(itemModel) as any);
-  }
-
-  updateItem(info: IUpdateItemPayload) {
-    this.store.dispatch(new item.UpdateItemAction(info) as any);
-  }
-
-  removeItem(i: IItem) {
-    this.store.dispatch(new item.RemoveItemAction(i) as any);
-  }
-
-  toggleCategoryExpanded(id: number) {
-    this._expanded = this._expanded.indexOf(id) > -1
-      ? this._expanded.filter(x => x !== id)
-      : [...this._expanded, id];
-    this.expanded$.next(this._expanded);
+  private logout() {
+    this.store.dispatch(new category.SecretPhraseChangeAction(undefined) as any);
+    this.store.dispatch({ type: 'USER_LOGOUT' });
+    this.router.navigate(['/']);
+    this.auth.logout();
   }
 }
