+++ src/app/content/content.module.ts
@@ -6,17 +6,21 @@ import { AuthHttp, AuthConfig } from 'angular2-jwt';
 import { ClipboardModule } from 'ngx-clipboard';
 import { NgxMessagesModule } from 'ngx-messages';
 
-import { reducer } from './store/reducers';
-import { CategoryEffects } from './store/effects/categories';
-import { ItemEffects } from './store/effects/items';
-
 import { ContentRoutingModule } from './content-routing.module';
+import { SharedModule } from '../shared/shared.module';
+
 import { ContentComponent } from './content.component';
-import { BackendService } from './services/backend.service';
 import { AddCategoryFormComponent } from './components/add-category-form/add-category-form.component';
 import { CategoryComponent, DeleteCategoryDialogComponent } from './components/category/category.component';
 import { ItemComponent, DeleteItemDialogComponent } from './components/item/item.component';
 import { ItemFormComponent } from './components/item-form/item-form.component';
+
+import { BackendService } from './services/backend.service';
+
+import { reducer } from './store/reducers';
+import { CategoryEffects } from './store/effects/categories';
+import { ItemEffects } from './store/effects/items';
+import { CategoryListComponent } from './components/category-list/category-list.component';
 
 export function authHttpServiceFactory(http: Http, options: RequestOptions) {
   return new AuthHttp(new AuthConfig({
@@ -31,6 +35,7 @@ export function authHttpServiceFactory(http: Http, options: RequestOptions) {
 @NgModule({
   imports: [
     ContentRoutingModule,
+    SharedModule,
     ClipboardModule,
     NgxMessagesModule,
     StoreModule.provideStore(reducer),
@@ -44,7 +49,8 @@ export function authHttpServiceFactory(http: Http, options: RequestOptions) {
     DeleteCategoryDialogComponent,
     ItemComponent,
     ItemFormComponent,
-    DeleteItemDialogComponent
+    DeleteItemDialogComponent,
+    CategoryListComponent
   ],
   providers: [
     BackendService,
